Moving the experiment to a more permanent location
