+++ src/altinn-app-frontend/src/shared/resources/utils/sagaSlice.ts
